[refactor] ASUtils renamed to AsyncStorageAPI, moved to API folder to better reflect role
